build.sh: Allow building external modules in parallel
Allow external modules to be built in parallel by letting projects
provide their own makefile for building external modules.

Currently, build.sh employs a simple for loop to build and install
modules listed in EXT_MODULES. Unfortunately, this prevents parallel
compilation across modules and results in poor utilization of build
machines with multiple CPUs.

If EXT_MODULES_MAKEFILE is set to the file path of a custom makefile,
build.sh will invoke make on it with all the necessary parameters that
are needed to build and install kernel modules.

Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: I357a524a831cf5a0a522fee66bc9c6afd249f1f1